[java] Enable and fix low SpotBugs warnings
This patch fixes the low SpotBugs issues along with some other minor
cleanup spotted along the way. Either the SpotBugs issues were fixed or
they were added to the excludeFilter as a special case.

Follow on patches will enforce SpotBugs in the pre-commit build.

Reviewed-on: http://gerrit.cloudera.org:8080/14801
Tested-by: Kudu Jenkins

+++ java/kudu-client/src/test/java/org/apache/kudu/client/TestOperation.java
@@ -81,12 +81,12 @@ public class TestOperation {
 
       // Check the strings.
       int offset = 3;
-      for (int i = 0; i <= 4; i++) {
+      for (long i = 0; i <= 4; i++) {
         // The offset into the indirect buffer
-        assertEquals(6 * i, Bytes.getLong(rows, offset));
+        assertEquals(6L * i, Bytes.getLong(rows, offset));
         offset += Longs.BYTES;
         // The length of the pointed-to string.
-        assertEquals(6, Bytes.getLong(rows, offset));
+        assertEquals(6L, Bytes.getLong(rows, offset));
         offset += Longs.BYTES;
       }
